Remove "PAUCAL"
This is an uber-simple dependency which adds genuinely deranged side-effects like headpatting. It also adds an easy supply chain attack by being hosted on some random dudes' Maven server with no file hashing and undisclosed administrators.

Also, to be perfectly clear--this commit is a lift and shift, nothing was rewritten.